Add telemetry processing and tiered billing

+++ core/src/main/java/com/powerpulse/core/telemetry/HomeLiveState.java
@@ -10,7 +10,7 @@ public record HomeLiveState(
         BigDecimal currentBillAmount,
         double quotaPercentage,
         int penaltyTier,
-        double penaltyMultiplier,
+        BigDecimal penaltyMultiplier,
         OffsetDateTime updatedAt
 ) {
 }
